Map scalars to colors when exporting slices as images
Fixes #1608

Works for both OrthogonalSlice and Slice.
If "Color Map Data" is selected in the slice property panel, in the
resulting output image scalars will be mapped to colors.

If "Color Map Data" is de-selected, the raw data will be written to
image (with potential conversion from float/double to unsigned char).

Opacity also works as expected for the image formats that support it.

As a result of this commit, the exported image will always end up
looking like the slice being drawn in the program.