Refactor histogram-creation background thread
This commit creates the histogram manager which can be used by non-UI
parts of the application to get the histogram for a dataset without
passing around pointers to parts of the UI.

We should avoid returning the image for the histogram until the
histogram has been completely computed.  Otherwise threading issues and
memory corruption occur.  Thus the histogram manager has an in-progress
state where getHistogram still returns nullptr but the histogram is
processing.