Simplified the set range dialog
There is now just one line for the range and the
standard OK and Cancel buttons.

Tooltips show the min and max values that can be set. The bottom
value in the range can never exceed the top value. And the top
value in the range can never be less than the bottom value.

The dialog box is fixed and cannot be resized.