Combined camera toolbar with utilities toolbar
This combines most actions from the camera toolbar with
the utilites toolbar. A few things were also done to clean
it up - such as removing the "Zoom to Data" button and creating
a drop-down menu for the "Reset view direction" buttons.

The CameraReaction class was copied from ParaView's "pqCameraReaction"
class with some modifications.