Combined axes toolbar with utilities toolbar
This combines the axes toolbar with the utilites toolbar,
and it re-writes a little bit of the axes toolbar functionality
to make it more specific to Tomviz. One thing that was changed is
that the "Reset Center" button acts on the bounds of the active
data source, and it does not gray out unless there is no data.