Some fixes to ensure the load state modal dialog gets the finished signal
There were some cases I missed where the signal could get missed both
for data sources with no operators and for operators that finish
extremely quickly.  This is basically some tweaks to when signals get
fired and when signal/slot connections are made that should prevent
the dialog from missing the signal in those cases.