Wait to close the state loading dialog until histograms finish
The volume module in 2d transfer function mode needs the histogram to be
complete before its state is fully loaded and user interaction during
the interval between the state loading and the histogram finishing can
cause its state to get messed up.  This delays closing the modal 'state
loading' dialog until the histograms are finished generating and the
volume module is in a consistent state.