Fix inconsistency between operator parameters in dialog/panel
Since the parameters can be modified not only in the panel but also in a
dedicated dialog, we should redraw the operator widget any time it is
selected in the pipeline, even if it was the current operator.

This ensures that the parameters in the panel are always in sync with
the values stored in the operator.