Update for changes to QVTKOpenGLWidget API changes
The signature of setEnableHiDPI was changed, and member variables
removed. The hack I was using to scale the DPI a little doesn't work in
the new class in VTK/ParaView master. This removes the override, but in
my testing suffers from fonts in the chart that look too small on a
Linux 4k high DPI display.