Read/write EMD data as row-major and convert for VTK
EMD files expect the data to be in row-major order and VTK expects
column major.  This commit flips the volume after it is read, which
allows the rest of the reader take the dimensions as they are in the
file and not flip them everywhere they are used.  This also will allow
other programs to read our EMD files without the data being flipped.