Better stability of python operators with child datasource
Modules attached to a child datasource would yield errors while the
underlying imagedata in uninitialized. This affects mostly the volume
and contour modules.

The errors would go away as soon as the operator sets progress data for
the first time or completes the transformation.

Solve this issue by making a shallow copy of the original datasource.