Reorder the EMD arrays for the volume
This is working on fixing issue #1808, where the code in this patch
ensures that a volume will be stored in the expected C order. I have
verified that this works as expected when loaded by matplotlib using the
simple Python example. Round tripping also worked with non-uniform
dimensions.

The major issue at this point is that all old EMDs will fail to load as
they were incorrectly stored in the Fortran order in the HDF5 file. I
also think that we need to treat image stacks differently in that the
slices should remain in contiguous memory so that they can be accessed.
This will require a special case for image stacks.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>