Allow a user to input negative angle increments
Fixes #1827, where if a user started with a positive
angle and ended with a negative angle, the data would
be all zeros.

In the `if()` statement in question, delta can be negative
under two conditions:

1. The start angle is **greater** than the end angle, and the start
image is **less** than the end image.
2. The start angle is **less** than the end angle, and the start
image is **greater** than the end image.

The bug in #1827 would appear via condition number 1. I believe
the `if()` block was designed only for condition number 2, so
this fixes the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>