Only stride in DataExchange read if any dim > 1100
Only stride in the Data Exchange reader if any dimension
is greater than 1100. Also added a message that prints
out to the terminal so that the user is aware of the stride,
and added a TODO comment that this should be a UI option
sooner or later.

I have now tested the Data Exchange reader and writer
by saving the file, opening it, saving it again, and then
comparing them with h5diff. h5diff reports them to be
identical.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>