Add generic HDF5 format reader
This reader first checks for the presence of a
`/exchange/data` dataset. If it exists, it assumes
it is a DataExchange file and will read it with the
DataExchangeFormat reader.

If that dataset does not exist, it finds all datasets
in the file that have 3 dimensions. If there is only
one dataset with three dimensions, it loads it automatically
as a volume. If there are multiple datasets with three
dimensions, it asks the user to pick one.

If the user tries to write an HDF5 file, it will just
write out a DataExchange format file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>