For EMD and DataExchange, use generic volume read
EMD, DataExchange, and GenericHDF5 all use the same
block of copy-pasted code for reading in volumes.

This commit reduces the amount of copy-pasted code
by calling the GenericHDF5Format function for reading
the volume.

However, as a consequence, EMD format will now also
stride when one of the dimensions is greater than 1100.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>