Turn off errors when checking if H5 dataset exists
It is acceptable for HDF5 errors to occur when checking if a dataset
exists, because the errors usually mean that it doesn't exist.

This commit prevents errors from printing when doing things such as
checking if `/exchange/data` is a dataset (which is currently
done to determine whether or not the file is of Data Exchange
format).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>