Give python readers precedence over json options
Currently, the two python readers that we have are for
".npy" files and ".mat" files. If a state file is loaded
that contains data from one of these two formats, it
attempts to read the files with ParaView (via the
"reader" in the json options) instead of with the
python readers.

Give the python readers precedence over the "reader"
in the json options to fix this issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>