Use the QVTKOpenGLNativeWidget as our base
We do not need anything specific from QVTKOpenGLWidget, and it is
causing intiialization/rendering issues on macOS right now. Move to the
QOpenGLWidget derived simpler class. Added a workaround to the event
handler to return true on mouse events and avoid passing them to higher
level widgets - this was causing an application window dragging bug on
KDE for me.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>