Improve isBuildDir to work for Linux and Windows
This was only really working for macOS, and as we are using this to
determine whether to set Python variables at runtime it is needed.
Thanks to Chris Harris for pointing it out, we still need to address the
case where we are not creating an app bundle with everything inside.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>