For HDF5 formats, allow reload and resample
For HDF5 formats, there is now a new context menu item for the
original data source in the pipeline view called "Reload and Resample".

If selected, it will prompt the user to choose a stride, and it
will then reload the dataset using the stride. It reloads the dataset
by writing over the previous vtkImageData. No new DataSource gets
created.

It seems to work for the most part, except for a few quirks:

1. The volume module disappears when it is reloaded and resampled.
2. The slice widget size does not update with the new volume size.
3. The camera does not seem to update with the new volume size.

I am looking through these. Let me know if there is a more preferable
way to do anything I've done so far.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>