Do not allow reload and resample for operated data
If a data source has operators (such as a square root operator),
not all of the data in the pipeline gets updated properly if
a reload and resample occurs. Only a few things get updated, such
as the units and size, and the color map.

For now, do not allow reload and resample for data with
operators, to avoid these issues.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>