Allow subvolume selection when reading HDF5 data
This adds to one of the h5cpp functions the ability to select a
subvolume when reading the HDF5 dataset. It does so by allowing
the caller to specify the start point for each dimension, and the
counts (number of points to read) for each dimension.

These variables get forwarded directly to HDF5's hyperslab function.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>