Add ability for user to select HDF5 subvolume
This adds a simple dialog that allows a user to select a subvolume and
a stride. The dialog also shows the user an estimated memory usage
(which is quite accurate) at the bottom.

When reading a volume from an HDF5 file for the first time, the dialog
will automatically appear if one of the dimensions is greater than
1200, so that the user can subsample if needed.

In addition, the dialog will appear if the user right-clicks the
dataset in the pipeline view and selects "Reload and Resample".

Note that the previous stride and subvolume choices are not saved,
and if the user chooses to "Reload and Resample", it will always
start with the default settings.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>