Allow reload, resample of HDF5 data with operators
Before this PR, reloading and resampling of HDF5 data that had
operators was not allowed, because the pipeline did not re-run.

This PR now allows for reloading and resampling of HDF5 data that
had operators, and it re-runs the pipeline if the data had operators.
It seems to work in my tests.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>