Fix SHIFT orientation before sending to python
The shift orientation was not being modified before it was being sent
to python. Modify the shift orientation to match the orientation that
the user selected.

This commit also places the tilt axis variable into the json
description. And it also ensures that the state of the orientation gets
re-loaded if the user double clicks on it in the pipeline after running.

The swapping of axes in python did not appear to be needed, so it was
removed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>