Save subsample settings and default to them
After subsampling the HDF5 file one time, save the subsample settings
and default to them the next time the subsampling dialog opens.

This keeps a record of the subsampling settings that were chosen in
the field data of the vtkImageData. This might potentially be used in
other features in the future as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>