EMD format can now write/read multiple scalars
When writing data to EMD format, if multiple scalar arrays are present
in the data, the active scalar array will be written to
"/data/tomography/data" (and given a "name" attribute with its scalar
array name), and all other scalar arrays will be written to a group
"/data/tomography/tomviz_scalars". Their dataset names are the same
as their scalar array names were.

When reading in an EMD file, tomviz checks to see if there is a group
in the EMD group named "tomviz_scalars". If there is, it assumes that
the datasets in the group are extra scalars to be read in, and it
reads them in and adds them.

All extra scalar arrays must have the same dimensions as the default
dataset. In addition, if any striding/subsampling was done on the
default dataset, it will also be performed on the exta scalar arrays.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>