Save subsample settings in the tomviz state file
This saves subsample settings in the tomviz state file, and it
reloads the data with the subsample settings if the user loads
the state file.

This commit also started using QJsonObject for passing options to
the different HDF5 formats, because it was getting a little
difficult to pass around and keep track of the various options. It
seems much easier with QJsonObject.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>