Ensure predicable scalar display order
The order of the scalars in the properties panel was being determined
by the pointer to the vtkDataArray, this meant that the order of the
scalars would change everytime a dataset was loaded. This was done
to ensure the the order didn't change when a scalar was renamed. To
ensure a more predictable order we now sort the scalars when a dataset
is selected and store that as the display order even after a rename.

Signed-off-by: Chris Harris <chris.harris@kitware.com>