Added optional help button to EditOperatorDialog
As of this commit, operators may specify whether they have a help
function or not by defining the virtual function `hasHelp()`. They
will then also define the virtual function `helpRequested()` to
define the behavior when help is requested.

If an operator has help, a help button is added to the
EditOperatorDialog in the bottom left corner. Upon pressing it, the
help button will run whatever the operator has specified to do.

This has been implemented in one case: the manual tilt axis alignment
operator. There is now a help button which opens up the web page on
"Read the Docs" which gives instructions for manual tilt axis alignment.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>