Utilize subsample settings in python pipeline
Use the subsample settings in the state file when reading EMD files in
the python pipeline. If there are multiple scalar arrays, it will use
the subsample settings on all of the arrays.

The memory usage has been tested, and it seems to use approximately the
expected amount of memory when subsampling a large dataset.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>