Require DataSource* for OperatorPython constructor
Before this commit, state files that contained python operators would
crash upon loading, because OperatorFactory::createOperator() would
create a python operator without a DataSource*. OperatorPython
frequently uses the parent DataSource*, so it ought to be required.

This commit adds it as a requirement, and ensures all of the python
operators are constructed with it. It also consequently fixes the
crashing bug.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>