Use function to prepend root web help path
The root web path for the help requests was removed, and it is now
added by a function instead. This is done so that, sometime in the
future, we may allow local paths to be used as well.

When we do this, basically, it will check local paths to see if the
file exists locally. If it doesn't exist, then it will open the web
page instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>