Fix integer overflows in Calculate2DHistogram()
For very large datasets (1501x1792x2048, for instance), there was at
least one overflow occuring in Calculate2DHistogram().

This commit attempts to avoid most overflows in the function by
converting one integer to a size_t for most cases where integers
are multiplied together and stored in a size_t.

This appears to have fixed the specific crash I was experiencing.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>