Read/write theta angles for DataExchangeFormat
This adds reading and writing for the theta angles in the
DataExchangeFormat. If the theta angles exist, the data will
be marked as a tilt series instead of as a volume, and the x
and z axes will be swapped so we can keep the tilt axis on
our fast axis.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>