Read tilt angles in data exchange
This reads in the theta angles from data exchange as tilt angles,
and it sets the tilt angles with them. Then, tomopy uses the tilt
angles instead of automatically generating them.

This also has the tomopy reconstruction script creating a child
dataset and returning the reconstruction results, because this is
needed to change it from a tilt series back to a volume.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>