Keep track of tilt axis in python pipeline
This adds a tilt axis variable so that we can keep track of the
tilt axis. It is especially important for tomopy, which wants the
tilt axis to be 0, and not 2.

This also adds an option to skip converting to Fortran ordering
when reading data exchange in the python pipeline. This could
potentially save a lot of time, because converting to Fortran is
unnecessary and may even slow down tomopy. In addition, tomopy
needs the tilt axis to be 0. The added option allows us to avoid
converting the tilt axis to 2 and then immediately back to 0 again.

The option is currently not in use anywhere, but it has been tested,
and it can be used in state files for pipelines.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>