Fix EMD dimension swapping in the executor script
Before, the way dimensions were swapped in the EMD reader was different
from the way dimensions were swapped in the EMD writer. This resulted
in "dim3" having the tilt angles instead of "dim1".

In the reader, the order of the dimensions were swapped. Whereas, in
the writer, the names of the dimensions were swapped.

This PR makes the swapping consistent so that the tilt angles are
written to the correct dimension.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>