Fix the "Delete Slices" operator
Before, in the EMD writer, the "dims1" dataset would be created
with the original size of the tilt_angles, and then the EMD writer
would attempt to re-assign the values. This would produce an error
if the size of the tilt_angles had changed (such as for operators
like "Delete Slices").

This fix gets "dims1" to be written with the new tilt_angles the
first time it is written, rather than through re-assignment.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>