Allow use of either transform_scalars or transform
With this change, the python script is searched for a transform()
function to use, and if one cannot be found, it then searches for a
transform_scalars() function to use instead. It will then use whichever
function it found, with preference given to transform().

In addition, the C++ code can distinguish whether transform() is being
used or transform_scalars(). This sets us up in a position where we
can pass transform() a different argument than transform_scalars(), but
they are currently being passed the same argument...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>