Check if custom operators re-implemented transform
This adds a proper check to see if custom operators have
re-implemented the `transform()` or `transform_scalars()`
functions that are in `Operator`. Before this change, the
base class `transform()` function would always be found in
classes that inherit from `Operator`, and they would attempt
to use that one, even though it is not implemented.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>