Added functions to convert to vtkDataObject
A new python function was added that tries to convert types that it
is given into a vtkDataObject. It can currently accept either a
vtkDataObject or the new DataObject class.

It is called in C++ to convert the Python::Object created from
reconstruction child datasets into a vtkDataObject. It works for
both the old transform_scalars() style and the new transform() style.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>