Fix reading of docker pipeline child datasets
Before this change, operators that produced child datasets and
did not update the progress would not produce any output. This
change fixes that issue.

It appears that `QDir::entryInfoList(QDir::NoDotAndDotDot)` does
not actually list files in the directory. Since we are only really
looking for files, it seems appropriate to use
`QDir::entryInfoList(QDir::Files)` instead. This results in the
child dataset being seen and read in.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>