Convert several similar itk operators to new API
These itk operators are now using the new API.

Note that a transpose is required at the end in order to put
the data back into Fortran indexing. This used to be done by
passing `isFortran=False` to the `tomviz.utils.set_array()`
function, but we cannot do that any longer because the results
are set as a property.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>