Convert Pad_Data to new API
The functionality of this is a little different now, because the
`minextent` variable is not being used. However, we can't use the
`minextent` variable in an external pipeline, so we might need to
remove it anyways.

This results in the whole volume shifting if any of the start paddings
are modified.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>