Set spacing on itk images, use itkutils function
This sets the spacing on itk images through the use of the new
dataset.spacing attribute. This was important in order for itk
images to perform some operations accurately.

This has been hidden in an itkutils function, however. So the
itkutils function is currently being used to generate the image
in most cases, except when a different datatype is desired.

All of the ITK operators were tested, and they now give the same
result with the new `transform()` API as they did with the old
`transform_scalars()` API.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>