Interpret external recon outputs as volumes
Before, the names of the dims would not be modified (and the contents
would not be swapped), so the first dim was seen to be "angles", and
the output was treated as a tilt series.

If the output does not have tilt angles, but the first dim is "angles",
then change the first dim to be "x" and the units to be "[n_m]" to
ensure the output is interpreted as a volume.

This also changes the dims to use a namedtuple to keep better track of
the contents.

Fixes: #1974

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>