Add decorators to automatically convert
These decorators, @with_vtk_dataobject and @with_dataset, are a simple
way for us to keep the functions backward compatible. They take the
first argument and automatically attempt to convert it to the desired
type before passing it on to the function. @with_vtk_dataobject also
checks to be sure that the application is running in internal mode.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>