In the external pipeline, do not allow float16
Although numpy supports float16, and it will automatically be used if
the data is small enough, C++ and VTK do not support float16. It is also
not one of the standard HDF5 types.

Make sure we do not write out float16 in the external pipeline. It is
not currently supported for reading in an HDF5 file, and it would
probably be difficult to support.

This incompatibility was manifested in crashing when the default custom
transform was performed in the external pipeline. That issue is now
fixed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>