Fix DockerUtilities tests
This includes putting in a 30 second time limit for most docker
commands (as opposed to the original 5 second time limit), and it
also includes sleeping before performing an inspect on a docker
image to give it time to get in its final state.

These changes were mostly important for the mac, as docker seems
to run pretty slowly on it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>