Increase mac attempts waiting for docker to start
Docker on the azure-pipelines mac images can be pretty slow in general.
It can also take a long time for it to start up. This can cause the
mac build to fail if waiting for docker to start takes too long.

Increase the max amount of time waiting for docker to start to
hopefully prevent more of these failures.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>