Offset plane bounds
Offset the plane bounds so that the plane does not get clipped and become
unsuable when it reaches the bounds of the data source. This fixes a bug that
would make the plane impossible to select by mouse if it was moved to the maximum
position in one of the orthogonal directions.

Signed-off-by: Brianna Major <brianna.major@kitware.com>