Fix plane visibility bugs
Toggling the module on and off would add a border to the clipping plane, but
toggling the plane visibility on and off would remove it. This removes the border
for all cases. Additionally, toggling the plane inversion would update the clip
filter, even when the module visibility was set to off. Only update clipping
plane when the clip module is visible.

Signed-off-by: Brianna Major <brianna.major@kitware.com>