Fix sample data generators
The decorator `@with_vtk_dataobject` was intended for use when the first
argument is a `dataobject` (and needs to be automatically converted to
a `vtkDataObject` type).

We could either fix this by re-ordering the arguments, or by just
removing the decorator. Since this function is probably only being
used internally, let's just remove the decorator.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>