Split xml string into many concatenated strings
Because of
[Compiler Error C2026](https://docs.microsoft.com/en-us/cpp/error-messages/compiler-errors-1/compiler-error-c2026?view=vs-2019)
on Windows, we cannot have a single c string exceed 16380 bytes. Our
xml has recently exceeded that, so we get this compiler error on
Windows.

This fixes the issue via their suggested solution: splitting the big
string up into many concatenated strings. It is done by replacing
newlines with two sets of double quotes, so that each line is an
independent c string.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>