Use python 3.8 for all azure-pipelines runs
Windows does not appear to have a library for python 3.7 available,
requiring us to move to python 3.8. We can move all of the other
operating systems forward as well, as long as we install python 3.8
h5py wheels from scipy (they are not available yet on PyPI), and we
skip installing pyfftw, as it does not have wheels either.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>