Clear tilt angles when type set to not tilt series
Some parts of the code rely on the presence of tilt angles to
determine the type of the data (for instance, [here](https://github.com/OpenChemistry/tomviz/blob/dd346564102c8efe37594b5675be2e950f43b788/tomviz/Pipeline.cxx#L376-L379)).

In this case, when the "Mark Data as Volume" operator set the
type to be a volume, it would not clear the tilt angles, and
Pipeline::branchFinished() would, immediately after, mark the
data as a tilt series due to the presence of the tilt angles.

To fix cases like this, clear the tilt angles when the type of
the data source is set to anything except a tilt series.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>