Use transformed data for the reconstruction widget
Previously, the reconstruction widget (which is used during a Simple
Back Projection (C++)) would use the parent data source information to
make its reconstruction image rather than the transformed data source
information.

This meant that, if an operator were performed that increased the
extents of the data (such as a Tilt Axis Alignment (Manual)), and
then a Simple Back Projection (C++) was performed, the reconstruction
image in the widget would have smaller extents than the transformed
data, and out-of-bounds writes would occur during updates, causing
weird reconstruction images and crashes.

Using the transformed data source in the reconstruction widget results
in the reconstruction image having the correct extents, and fixes the
issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>