Add options to view dark and white data
This adds two menu items under "View": "Show Dark Data" and
"Show White Data". If the currently activated data source
contains dark and white data, these two menu items will be
enabled. Selecting the menu items results in a blocking
pop-up window that shows an image slice of the dark/white data.
The image slice uses the same color map as the data source.
The interactor is currently vtkInteractorStyleRubberBand2D, just
because that is what was used for the image slices in the
RotateAlignWidget.

This currently views the dark/white data as a slice along the
z axis, which might not always be the case.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>