Add SliceViewDialog for viewing dark/white data
This adds a small dialog with radio buttons for switching to dark and
white data. The image shows the currently selected one.

More work needs to be done. We need to rescale the color map so that
the dark data can be visualized more easily. We should also add the
ability to move the mouse over pixels and, somewhere, the pixel value
will be shown.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>