Rescale lookup table points for dark/white viewing
This copies over some functions from ParaView, which are used to
rescale the lookup table points for the color map of the dark/white
data viewer.

This rescales the color map lookup table points to be within the data
range so that the data (the dark data, especially, since it normally
has a small data range) is more easily viewable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>