Add ability to load multiple volumes from HDF5
When there are multiple volumes in an HDF5 file, instead of making the
user pick one, allow the user to pick multiple. This is done via a
list of checkboxes.

If the user selects multiple volumes, they will all be added to the
same image data as additional scalar arrays. The active array may
be changed via the properties panel.

Additionally, an extra safety check was added to make sure that the
additional scalar arrays have the same dimensions (or else crashing
will occur).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>