Never ask for subsampling during state file load
When a state file is loaded, never prompt the user for subsampling.
This fixes numerous cases where the subsampling prompt is undesirable,
including loading state files via the python console.

If the user wishes to change the subsampling of a state file, they need
to modify the state file.

Additionally, this ensures that when the subsample volume bounds are
negative, the final volume bounds get saved to the image, and will be
written to the state file as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>