Allow subsample settings use in interactive python
When a data source is constructed from a set of files, allow the user
to pass in a "subsampleSettings" dict which contains subsampling
options. These subsample settings will only be currently used for HDF5
type files.

The dict may contain a "strides" key which would have a value of a list
of ints of size 3 (one for each dimension), and the dict may contain a
"volumeBounds" key, which would have a list of ints of size 6 (for
dimension i, index i * 2 is the lower bound, and index i * 2 + 1 is the
upper bound).

This saves the subsample settings in the "reader" of the data source
json object. To allow for this to work, we also now check for the
subsample settings in the "reader" in ModuleManager::loadDataSource().

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>