The code to write DataExchange got deleted
It appears that the code to write the correct format based on the ending
was deleted, and the current behavior was just to change the file
ending. Add back some logic to select the correct file format based upon
the ending determined. This should correct issues I was seeing where the
dark/white data was not getting through to the external Python pipeline.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>