Use regular expressions to match trailing newlines
This is a much more reliable way than the previous method
to eliminate trailing newlines/carriage returns. The regular
expression should only match to a carriage return that is trailing,
and it will only match with either the unix return ('\n') or the
windows return ('\r\n'). If the line ends with either of
these types of returns, the carriage return will be removed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>