Save active scalars as soft link in tomviz_scalars
In the EMD format in tomviz, previously, the active scalars would be
written to "/data/tomography/data", and they would not be written
anywhere else. Any extra scalars were written to
"/data/tomography/tomviz_scalars". Now, a soft link to the active
scalars is also created in "/data/tomography/tomviz_scalars", and
it points to "/data/tomography/data". This is done so that users
can get all the scalars in "/data/tomography/tomviz_scalars" if
they would like.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>