Convert array back to fortran before setting
Unfortunately, it looks like numpy.ndarray.swapaxes converts
the ordering from Fortran to C. When this happens, a warning
message will be printed to the console later that a conversion to
Fortran ordering is required.

Instead of printing the warning, just convert it back to Fortran
ordering ourselves. We already know that this operator unfortunately
converts the ordering to C.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>