Always use component 0 when getting voxel values
The component was previously being used to select the active scalars
before getting the voxel value. However,
vtkImageData::GetScalarComponentAsDouble() already uses the active
scalars.

Additionally, most of our data only has one component. Thus, get the
value from component 0. This prevents an error when moving the mouse
over a slice of data that does not have a scalars index of 0 (which
will happen if you have a data set loaded with multiple scalars, and
you are not visualizing the first set of scalars).

An example of the error that is fixed is below:
```
(  16.656s) [paraview        ]       vtkImageData.cxx:1333   ERR| vtkImageData (0x56254c421da0): Bad component index 1
```

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>